tools/libs: usuobj: refine method generate_uobj_info to generate ustack and tstack TOS

+++ src/include/uberspark.h
@@ -78,8 +78,8 @@ typedef void * uobj_entrystub_t;
 
 typedef struct {
 	uobj_entrystub_t entrystub;
-	uint8_t ustack_tos[MAX_PLATFORM_CPUS];
-	uint8_t tstack_tos[MAX_PLATFORM_CPUS];
+	uint32_t ustack_tos[MAX_PLATFORM_CPUS];
+	uint32_t tstack_tos[MAX_PLATFORM_CPUS];
 } __attribute__((packed)) uobj_info_t;
 
 typedef struct {
